refer/testing2/requirements.txt
llama-index-core~=0.12.17
pillow~=11.1.0
python-pptx~=1.0.2
pymilvus~=2.5.4
vanna~=0.7.6
chainlit~=2.0.5
starlette~=0.37.2
pydantic~=2.10.2
openai~=1.64.0
python-dotenv~=1.0.1
uvicorn~=0.30.6
fastapi~=0.111.0
Flask~=3.1.0
aiofiles~=23.2.1
aiohttp~=3.11.11
literalai~=0.1.103
SQLAlchemy~=2.0.37
minio~=7.2.15
PyMySQL~=1.1.1
aiomysql~=0.2.0
docling~=2.15.1
pandas~=2.2.3
huggingface-hub~=0.27.1
plotly~=6.0.0
requests~=2.32.3
sqlparse~=0.5.3
httpx~=0.27.0
autogen-agentchat
autogen-ext[openai]

pytest~=8.3.5
allure-python-commons~=2.13.5
PyJWT~=2.10.1
passlib~=1.7.4
DBUtils~=3.1.0
PyYAML~=6.0.2
loguru~=0.7.2
aerich~=0.7.2
pydantic-settings~=2.4.0